Refactor: API 매퍼 분리 #105
