+++ lib/features/user/data/repositories/user_repository.dart
@@ -2,6 +2,7 @@ import 'package:commonplant_frontend/core/network/api_client.dart';
 import 'package:commonplant_frontend/core/network/api_response_parser.dart';
 import 'package:commonplant_frontend/features/user/data/datasources/user_remote_data_source.dart';
 import 'package:commonplant_frontend/features/user/data/dtos/user_requests.dart';
+import 'package:commonplant_frontend/features/user/data/mappers/user_mapper.dart';
 import 'package:commonplant_frontend/features/user/domain/entities/user_profile.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -46,15 +47,4 @@ class UserRepository {
   Future<void> deleteMe() async {
     await _remoteDataSource.deleteMe();
   }
-}
-
-UserProfile userProfileFromJson(JsonMap json) {
-  return UserProfile(
-    id: readRequiredString(json, const ['id', 'userId', 'nanoId'], '사용자 ID'),
-    name: readRequiredString(json, const ['name'], '사용자 이름'),
-    email: readOptionalString(json, const ['email']),
-    provider: readOptionalString(json, const ['provider']),
-    imgUrl: readOptionalString(json, const ['imgUrl', 'imageUrl']),
-    introduction: readOptionalString(json, const ['introduction']),
-  );
 }
